Reverted parameters change, renamed an sns topic
Moved email addresses back to hardcoded text in the template, as couldn't find way to provide them
as parameters, other than passing them in via cloudFormation deploy parameters, which meant hardcoding
them in a script instead of the template.
Renamed one of the SNS topics so reason for each topic is more clear - results is for the normal messages,
errors is for errors.